# management-api-server/doc/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "This is a Restful service for operating Apache Cassandra.  You can find out more about the Management API on [Github](http://github.com/k8ssandra/management-api-for-apache-cassandra)",
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "title" : "Management API for Apache Cassandra™",
    "version" : "0.1"
  },
  "paths" : {
...
    "/api/v0/ops/tables" : {
      "get" : {
        "operationId" : "listTables",
        "parameters" : [ {
          "in" : "query",
          "name" : "keyspaceName",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "example" : [ "table_1", "table_2" ],
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "type" : "string"
                  }
                }
              }
            },
            "description" : "Table list"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "List tables failed. Non-empty 'keyspaceName' must be provided"
              }
            },
            "description" : "Keyspace name not provided"
          }
        },
        "summary" : "List the table names in the given keyspace"
      }
    },
    "/api/v0/ops/tables/compact" : {
      "post" : {
        "operationId" : "compact",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompactRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "Table compaction successful"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "Invalid option combination: Can not use split-output here"
              }
            },
            "description" : "Invalid table compaction request"
          }
        },
        "summary" : "Force a (major) compaction on one or more tables or user-defined compaction on given SSTables"
      }
    },
    "/api/v0/ops/tables/create" : {
      "post" : {
        "operationId" : "createTable",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTableRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "Table creation successful"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "Table creation failed: some failure message"
              }
            },
            "description" : "Table creation failed"
          }
        },
        "summary" : "Create a new table in an existing keyspace"
      }
    },
    "/api/v0/ops/tables/flush" : {
      "post" : {
        "operationId" : "flush",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "Table flush successful"
          }
        },
        "summary" : "Flush one or more tables"
      }
    },
    "/api/v0/ops/tables/garbagecollect" : {
      "post" : {
        "operationId" : "garbageCollect",
        "parameters" : [ {
          "in" : "query",
          "name" : "tombstoneOption",
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "Table garbage collection successful"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "tombstoneOption must be either ROW or CELL"
              }
            },
            "description" : "Invalid table garbage collection request"
          }
        },
        "summary" : "Remove deleted data from one or more tables"
      }
    },
    "/api/v0/ops/tables/scrub" : {
      "post" : {
        "operationId" : "scrub",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ScrubRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "Table scrub successful"
          }
        },
        "summary" : "Scrub (rebuild sstables for) one or more tables"
      }
    },
    "/api/v0/ops/tables/sstables/upgrade" : {
      "post" : {
        "operationId" : "upgradeSSTables",
        "parameters" : [ {
          "in" : "query",
          "name" : "excludeCurrentVersion",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "text/plain" : {
                "example" : "OK"
              }
            },
            "description" : "SSTable upgrade successful"
          }
        },
        "summary" : "Rewrite sstables (for the requested tables) that are not on the current version (thus upgrading them to said current version)"
      }
    },
...
    "/api/v1/ops/tables/compact" : {
      "post" : {
        "operationId" : "compact_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CompactRequest"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "content" : {
              "text/plain" : {
                "example" : "d69d1d95-9348-4460-95d2-ae342870fade",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for keyspace compaction process"
          },
          "400" : {
            "content" : {
              "text/plain" : {
                "example" : "Invalid option combination: Can not use split-output here"
              }
            },
            "description" : "Invalid table compaction request"
          }
        },
        "summary" : "Force a (major) compaction on one or more tables or user-defined compaction on given SSTables"
      }
    },
    "/api/v1/ops/tables/scrub" : {
      "post" : {
        "operationId" : "scrub_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ScrubRequest"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "content" : {
              "text/plain" : {
                "example" : "d69d1d95-9348-4460-95d2-ae342870fade",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for table scrub process"
          }
        },
        "summary" : "Scrub (rebuild sstables for) one or more tables. This operation is asynchronous and returns immediately."
      }
    },
    "/api/v1/ops/tables/sstables/upgrade" : {
      "post" : {
        "operationId" : "upgradeSSTables_1",
        "parameters" : [ {
          "in" : "query",
          "name" : "excludeCurrentVersion",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/KeyspaceRequest"
              }
            }
          }
        },
        "responses" : {
          "202" : {
            "content" : {
              "text/plain" : {
                "example" : "d69d1d95-9348-4460-95d2-ae342870fade",
                "schema" : {
                  "type" : "string"
                }
              }
            },
            "description" : "Job ID for keyspace SSTable upgrade process"
          }
        },
        "summary" : "Rewrite sstables (for the requested tables) that are not on the current version (thus upgrading them to said current version). This operation is asynchronous and returns immediately."
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "Column" : {
        "type" : "object",
        "properties" : {
          "kind" : {
            "type" : "string",
            "enum" : [ "PARTITION_KEY", "CLUSTERING_COLUMN", "REGULAR", "STATIC" ]
          },
          "name" : {
            "type" : "string"
          },
          "order" : {
            "type" : "string",
            "enum" : [ "ASC", "DESC" ]
          },
          "position" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string"
          }
        },
        "required" : [ "name", "type" ]
      },
      "CompactRequest" : {
        "type" : "object",
        "properties" : {
          "end_token" : {
            "type" : "string"
          },
          "keyspace_name" : {
            "type" : "string"
          },
          "split_output" : {
            "type" : "boolean"
          },
          "start_token" : {
            "type" : "string"
          },
          "tables" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          },
          "user_defined" : {
            "type" : "boolean"
          },
          "user_defined_files" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "keyspace_name", "split_output", "user_defined" ]
      },
...
      "CreateTableRequest" : {
        "type" : "object",
        "properties" : {
          "columns" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Column"
            }
          },
          "keyspace_name" : {
            "type" : "string"
          },
          "options" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          },
          "table_name" : {
            "type" : "string"
          }
        },
        "required" : [ "columns", "keyspace_name", "table_name" ]
      },
...
      "KeyspaceRequest" : {
        "type" : "object",
        "properties" : {
          "jobs" : {
            "type" : "integer",
            "format" : "int32"
          },
          "keyspace_name" : {
            "type" : "string"
          },
          "tables" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "jobs", "keyspace_name", "tables" ]
      },
...
      "ScrubRequest" : {
        "type" : "object",
        "properties" : {
          "check_data" : {
            "type" : "boolean"
          },
          "disable_snapshot" : {
            "type" : "boolean"
          },
          "jobs" : {
            "type" : "integer",
            "format" : "int32"
          },
          "keyspace_name" : {
            "type" : "string"
          },
          "reinsert_overflowed_ttl" : {
            "type" : "boolean"
          },
          "skip_corrupted" : {
            "type" : "boolean"
          },
          "tables" : {
            "type" : "array",
            "items" : {
              "type" : "string"
            }
          }
        },
        "required" : [ "check_data", "disable_snapshot", "jobs", "keyspace_name", "reinsert_overflowed_ttl", "skip_corrupted", "tables" ]
      },
...
    }
  }
}